Backend API/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
numpy==1.25.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
python-multipart==0.0.6
pydantic==2.5.0
aiofiles==23.2.1
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
ollama==0.1.7
